move removed kind evaluation to genericapiserver
